Allow configuration of AvroBytesDecoder when using web console to load data
Unlike `avro_ocf` which doesn't require any configuration the
`avro_stream` input format requires the user to either provide schema(s)
or a mechanism to resolve the schema for a message.

This changes adds configuration options to the input format phase to
allow specifying these options in a structured way and allow override
similar to how Kafka consumerProperties is handled.